pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Eraldo Energy <eraldo@eraldo.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
strawberry-graphql-django = "^0.41.1"
strawberry-graphql = "^0.232.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
